+++ bin/requirements.txt
@@ -1,6 +1,6 @@
 biopython==1.79
 pysam==0.17.0
-dataclasses==0.8
+dataclasses==0.6
 pandas==1.1.5
 cyvcf2==0.30.12
 gtfparse==1.2.1
