+++ tests/scripts/test_00_help.sh
