+++ tests/scripts/test_00_test_mode.sh
