+++ tests/scripts/test_03.sh
@@ -2,16 +2,18 @@
 
 ##################################################################################
 # FASTQ input
-# single-end reads
-# use custom reference
+# paired-end reads
+# --keep_intermediate has BAM files in output
 ##################################################################################
-echo "Running CoVigator pipeline test 10"
+echo "Running CoVigator pipeline test 3"
 source bin/assert.sh
-output=output/test10
-echo -e "test_data\t"`pwd`"/test_data/test_data_1.fastq.gz\n" > test_data/test_input.txt
-nextflow main.nf -profile test,conda --input_fastqs_list test_data/test_input.txt \
---library single --output $output \
---reference `pwd`/reference/Sars_cov_2.ASM985889v3.dna.toplevel.fa
+output=tests/output/test3
+nextflow main.nf -profile test,conda --name test_data \
+--output $output \
+--fastq1 tests/test_data/test_data_1.fastq.gz \
+--fastq2 tests/test_data/test_data_2.fastq.gz \
+--keep_intermediate \
+--skip_ivar --skip_bcftools --skip_gatk
 
 test -s $output/test_data.lofreq.vcf.gz || { echo "Missing VCF output file!"; exit 1; }
 test -s $output/test_data.fastp_stats.json || { echo "Missing VCF output file!"; exit 1; }
@@ -20,7 +22,11 @@ test -s $output/test_data.coverage.tsv || { echo "Missing coverage output file!"
 test -s $output/test_data.depth.tsv || { echo "Missing depth output file!"; exit 1; }
 test -s $output/test_data.lofreq.pangolin.csv || { echo "Missing pangolin output file!"; exit 1; }
 
-assert_eq `zcat $output/test_data.lofreq.vcf.gz | grep -v '#' | wc -l` 11 "Wrong number of variants"
+# these are the intermediate files kept by --keep_intermediate
+test -s $output/test_data.preprocessed.bam || { echo "Missing BAM file!"; exit 1; }
+test -s $output/test_data.preprocessed.bai || { echo "Missing BAI file!"; exit 1; }
+
+assert_eq `zcat $output/test_data.lofreq.vcf.gz | grep -v '#' | wc -l` 54 "Wrong number of variants"
 assert_eq `zcat $output/test_data.lofreq.vcf.gz | grep -v '#' | grep PASS | wc -l` 2 "Wrong number of variants"
 
 assert_eq `cat $output/test_data.lofreq.pangolin.csv |  wc -l` 2 "Wrong number of pangolin results"
