+++ tests/scripts/test_08.sh
@@ -1,15 +1,18 @@
 #!/bin/bash
 
 ##################################################################################
-# primer trimmming
+# FASTQ input
+# paired-end reads
 ##################################################################################
-echo "Running CoVigator pipeline test 11"
+echo "Running CoVigator pipeline test 8"
 source bin/assert.sh
-output=output/test10
-echo -e "test_data\t"`pwd`"/test_data/test_data_1.fastq.gz\n" > test_data/test_input.txt
-nextflow main.nf -profile test,conda --input_fastqs_list test_data/test_input.txt \
---library single --output $output \
---primers test_data/SARS-CoV-2.primer.bed
+output=tests/output/test8
+nextflow main.nf -profile test,conda --name test_data \
+--output $output \
+--fastq1 tests/test_data/test_data_1.fastq.gz \
+--fastq2 tests/test_data/test_data_2.fastq.gz --skip_sarscov2_annotations \
+--keep_intermediate \
+--skip_ivar --skip_bcftools --skip_gatk
 
 test -s $output/test_data.lofreq.vcf.gz || { echo "Missing VCF output file!"; exit 1; }
 test -s $output/test_data.fastp_stats.json || { echo "Missing VCF output file!"; exit 1; }
@@ -18,7 +21,8 @@ test -s $output/test_data.coverage.tsv || { echo "Missing coverage output file!"
 test -s $output/test_data.depth.tsv || { echo "Missing depth output file!"; exit 1; }
 test -s $output/test_data.lofreq.pangolin.csv || { echo "Missing pangolin output file!"; exit 1; }
 
-assert_eq `zcat $output/test_data.lofreq.vcf.gz | grep -v '#' | wc -l` 3 "Wrong number of variants"
-assert_eq `zcat $output/test_data.lofreq.vcf.gz | grep -v '#' | grep PASS | wc -l` 0 "Wrong number of variants"
+assert_eq `zcat $output/test_data.lofreq.vcf.gz | grep -v '#' | wc -l` 54 "Wrong number of variants"
+assert_eq `zcat $output/test_data.lofreq.vcf.gz | grep -v '#' | grep PASS | wc -l` 2 "Wrong number of variants"
+assert_eq `zcat $output/test_data.lofreq.vcf.gz | grep -v '#' | grep PFAM | wc -l` 0 "Wrong number of variants"
 
 assert_eq `cat $output/test_data.lofreq.pangolin.csv |  wc -l` 2 "Wrong number of pangolin results"
