+++ tests/scripts/test_09.sh
@@ -2,16 +2,16 @@
 
 ##################################################################################
 # FASTQ input
-# paired-end reads
-# use --input_fastqs_list
+# single-end reads
+# use custom reference
 ##################################################################################
-echo "Running CoVigator pipeline test 6"
+echo "Running CoVigator pipeline test 09"
 source bin/assert.sh
-output=output/test6
-echo -e "test_data\t"`pwd`"/test_data/test_data_1.fastq.gz\t"`pwd`"/test_data/test_data_2.fastq.gz\n" > test_data/test_input.txt
-nextflow main.nf -profile test,conda --input_fastqs_list test_data/test_input.txt \
---library paired --output $output \
---skip_ivar --skip_bcftools --skip_gatk
+output=tests/output/test09
+echo -e "test_data\t"`pwd`"/tests/test_data/test_data_1.fastq.gz\n" > tests/test_data/test_input.txt
+nextflow main.nf -profile test,conda --input_fastqs_list tests/test_data/test_input.txt \
+--library single --output $output \
+--reference `pwd`/reference/Sars_cov_2.ASM985889v3.dna.toplevel.fa
 
 test -s $output/test_data.lofreq.vcf.gz || { echo "Missing VCF output file!"; exit 1; }
 test -s $output/test_data.fastp_stats.json || { echo "Missing VCF output file!"; exit 1; }
@@ -20,7 +20,7 @@ test -s $output/test_data.coverage.tsv || { echo "Missing coverage output file!"
 test -s $output/test_data.depth.tsv || { echo "Missing depth output file!"; exit 1; }
 test -s $output/test_data.lofreq.pangolin.csv || { echo "Missing pangolin output file!"; exit 1; }
 
-assert_eq `zcat $output/test_data.lofreq.vcf.gz | grep -v '#' | wc -l` 54 "Wrong number of variants"
+assert_eq `zcat $output/test_data.lofreq.vcf.gz | grep -v '#' | wc -l` 11 "Wrong number of variants"
 assert_eq `zcat $output/test_data.lofreq.vcf.gz | grep -v '#' | grep PASS | wc -l` 2 "Wrong number of variants"
 
 assert_eq `cat $output/test_data.lofreq.pangolin.csv |  wc -l` 2 "Wrong number of pangolin results"
