+++ tests/scripts/test_10.sh
@@ -1,17 +1,15 @@
 #!/bin/bash
 
 ##################################################################################
-# FASTQ input
-# single-end reads
-# --keep_intermediate has BAM files in output
+# primer trimmming
 ##################################################################################
-echo "Running CoVigator pipeline test 2"
+echo "Running CoVigator pipeline test 10"
 source bin/assert.sh
-output=output/test2
-nextflow main.nf -profile test,conda --name test_data \
---output $output \
---fastq1 test_data/test_data_1.fastq.gz --keep_intermediate \
---skip_ivar --skip_bcftools --skip_gatk
+output=tests/output/test10
+echo -e "test_data\t"`pwd`"/tests/test_data/test_data_1.fastq.gz\n" > tests/test_data/test_input.txt
+nextflow main.nf -profile test,conda --input_fastqs_list tests/test_data/test_input.txt \
+--library single --output $output \
+--primers tests/test_data/SARS-CoV-2.primer.bed
 
 test -s $output/test_data.lofreq.vcf.gz || { echo "Missing VCF output file!"; exit 1; }
 test -s $output/test_data.fastp_stats.json || { echo "Missing VCF output file!"; exit 1; }
@@ -20,11 +18,7 @@ test -s $output/test_data.coverage.tsv || { echo "Missing coverage output file!"
 test -s $output/test_data.depth.tsv || { echo "Missing depth output file!"; exit 1; }
 test -s $output/test_data.lofreq.pangolin.csv || { echo "Missing pangolin output file!"; exit 1; }
 
-# these are the intermediate files kept by --keep_intermediate
-test -s $output/test_data.preprocessed.bam || { echo "Missing BAM file!"; exit 1; }
-test -s $output/test_data.preprocessed.bai || { echo "Missing BAI file!"; exit 1; }
-
-assert_eq `zcat $output/test_data.lofreq.vcf.gz | grep -v '#' | wc -l` 11 "Wrong number of variants"
-assert_eq `zcat $output/test_data.lofreq.vcf.gz | grep -v '#' | grep PASS | wc -l` 2 "Wrong number of variants"
+assert_eq `zcat $output/test_data.lofreq.vcf.gz | grep -v '#' | wc -l` 3 "Wrong number of variants"
+assert_eq `zcat $output/test_data.lofreq.vcf.gz | grep -v '#' | grep PASS | wc -l` 0 "Wrong number of variants"
 
 assert_eq `cat $output/test_data.lofreq.pangolin.csv |  wc -l` 2 "Wrong number of pangolin results"
