+++ tests/test_data/SARS-CoV-2.primer.bed
